requirements.txt
click~=7.1.2
pyfiglet~=0.8.post1
setuptools~=50.3.2
pandas~=1.1.4
pydantic~=1.7.2
arrow~=0.17.0
PyMySQL~=0.10.1
Jinja2~=2.11.2
mysql-connector-python~=8.0.22
etl~=0.1